feat!: update to the 2023-09 job template schema (#4)
The job template schema is being updated for service release. With it,
the implementation library is being updated in a breaking way and so
this code needs to be updated to that revision. This includes a rename
of the job template's specification name from the codename openjobio
to Open Job Description.

Signed-off-by: Daniel Neilson <neilsd@amazon.com>
Signed-off-by: Mark Wiebe <markw@amazon.com>