feat: Apply path mapping for OCIO configs with absolute search paths (#34)
In the NukeClient, determine if a custom OCIO config is being used, and if it contains absolute search paths. If so, create a temporary copy of the config where all of the paths are absolute and have had path mapping applied. Since the temporary file is being stored in a temporary directory (in the case that the original config's directory is read-only), the the relative paths will not be valid.

---------

Signed-off-by: Alex Gerveshi <15710060+agerveshi@users.noreply.github.com>